# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Playlisted API
  version: 0.2.0
  summary: OpenAPI-first REST API for the Playlisted audio platform.
  description: |
    Contract for the Playlisted backend. The server validates requests and responses
    against this document, and the client SDK is generated from the same source.
servers:
  - url: http://localhost:4000
    description: Local development
security: []
...
paths:
...
  /api/v1/me/playlists:
    get:
      tags: [Me]
      operationId: listMyPlaylists
      summary: List playlists owned by the current user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Owned playlists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/me/collections/playlists:
    get:
      tags: [Me]
      operationId: listCollectionPlaylists
      summary: List playlists saved to the current user's collections
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Saved collection playlists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionPlaylistsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/me/collections/playlists/{playlistId}:
    post:
      tags: [Me]
      operationId: addCollectionPlaylist
      summary: Add a playlist to the current user's collections
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: playlistId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Already owned or already saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionPlaylistSavedResponse'
        '201':
          description: Saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionPlaylistSavedResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Playlist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Me]
      operationId: removeCollectionPlaylist
      summary: Remove a playlist from the current user's collections
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: playlistId
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Removed
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/me/favorites/playlists:
    get:
      tags: [Me]
      operationId: listFavoritePlaylists
      summary: List the current user's favorited playlists
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Favorite playlists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoritePlaylistsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/me/favorites/playlists/{playlistId}:
    post:
      tags: [Me]
      operationId: addFavoritePlaylist
      summary: Add a playlist to favorites
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: playlistId
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoritePlaylistSavedResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Playlist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Me]
      operationId: removeFavoritePlaylist
      summary: Remove a playlist from favorites
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: playlistId
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Removed
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/users/by-username/{username}/playlists/{slug}:
    get:
      tags: [Users]
      operationId: getPlaylistByUsernameAndSlug
      summary: Get playlist detail by owner username and slug
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Playlist detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistDetail'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User or playlist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/playlists:
    get:
      tags: [Playlists]
      operationId: listPlaylists
      summary: List playlists
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: ownerId
          schema:
            type: string
        - in: query
          name: visibility
          schema:
            $ref: '#/components/schemas/Visibility'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/PublishStatus'
      responses:
        '200':
          description: Playlist collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistListResponse'
    post:
      tags: [Playlists]
      operationId: createPlaylist
      summary: Create a playlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlaylistRequest'
      responses:
        '201':
          description: Playlist created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistDetail'
        '404':
          description: Owner not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Slug already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/playlists/{playlistId}:
    get:
      tags: [Playlists]
      operationId: getPlaylistById
      summary: Get playlist detail
      parameters:
        - in: path
          name: playlistId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Playlist detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistDetail'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Playlist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Playlists]
      operationId: deletePlaylist
      summary: Delete a playlist owned by the current user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: playlistId
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Playlist deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Playlist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [Playlists]
      operationId: updatePlaylist
      summary: Update a playlist owned by the current user
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: playlistId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePlaylistRequest'
      responses:
        '200':
          description: Updated playlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistDetail'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Playlist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/playlists/{playlistId}/items:
    post:
      tags: [Playlists]
      operationId: addPlaylistItem
      summary: Add a recording to a playlist
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: playlistId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddPlaylistItemRequest'
      responses:
        '200':
          description: Updated playlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistDetail'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Playlist or recording not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Item already in playlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/playlists/{playlistId}/items/{recordingId}:
    delete:
      tags: [Playlists]
      operationId: removePlaylistItem
      summary: Remove a recording from a playlist
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: playlistId
          required: true
          schema:
            type: string
        - in: path
          name: recordingId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated playlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistDetail'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Playlist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/playlists/{playlistId}/items/reorder:
    put:
      tags: [Playlists]
      operationId: reorderPlaylistItems
      summary: Reorder tracks in a playlist
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: playlistId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderPlaylistItemsRequest'
      responses:
        '200':
          description: Updated playlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistDetail'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Playlist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/playlists/{playlistId}/tags:
    put:
      tags: [Playlists]
      operationId: setPlaylistTags
      summary: Replace all tag associations on a playlist
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: playlistId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetPlaylistTagsRequest'
      responses:
        '200':
          description: Updated playlist with new tags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistDetail'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Playlist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: pageSize
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    PlaylistType:
      type: string
      enum: [PLAYLIST, ALBUM, MIX, PODCAST_CHANNEL, RELEASE]
...
    Visibility:
      type: string
      enum: [PUBLIC, UNLISTED, PRIVATE]
    PublishStatus:
      type: string
      enum: [DRAFT, PUBLISHED, ARCHIVED]
...
    ErrorResponse:
      type: object
      additionalProperties: true
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details: {}
    PaginationMeta:
      type: object
      additionalProperties: false
      required: [page, pageSize, total]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
...
    PlaylistDetail:
      type: object
      additionalProperties: false
      required:
        [id, href, ownerId, title, slug, type, visibility, status, featured, isPinnedOnProfile, itemCount, totalDurationSeconds, createdAt, updatedAt, owner, recordings]
      properties:
        id:
          type: string
        href:
          type: string
        ownerId:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: [string, 'null']
        coverArtUrl:
          type: [string, 'null']
          format: uri-reference
        type:
          $ref: '#/components/schemas/PlaylistType'
        visibility:
          $ref: '#/components/schemas/Visibility'
        status:
          $ref: '#/components/schemas/PublishStatus'
        featured:
          type: boolean
        isPinnedOnProfile:
          type: boolean
        itemCount:
          type: integer
        totalDurationSeconds:
          type: integer
        publishedAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        owner:
          $ref: '#/components/schemas/OwnerSummary'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        recordings:
          type: array
          items:
            $ref: '#/components/schemas/RecordingInPlaylist'
    PlaylistListResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PlaylistSummary'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    CreatePlaylistRequest:
      type: object
      additionalProperties: false
      required: [ownerId, title]
      properties:
        ownerId:
          type: string
        title:
          type: string
          minLength: 1
          maxLength: 191
        description:
          type: [string, 'null']
        coverArtUrl:
          type: [string, 'null']
          format: uri-reference
        type:
          $ref: '#/components/schemas/PlaylistType'
        visibility:
          $ref: '#/components/schemas/Visibility'
        status:
          $ref: '#/components/schemas/PublishStatus'
        publishedAt:
          type: [string, 'null']
          format: date-time
...
    UpdatePlaylistRequest:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 191
        description:
          type: [string, 'null']
        coverArtUrl:
          type: [string, 'null']
          format: uri-reference
        type:
          $ref: '#/components/schemas/PlaylistType'
        visibility:
          $ref: '#/components/schemas/Visibility'
        status:
          $ref: '#/components/schemas/PublishStatus'
        isPinnedOnProfile:
          type: boolean
        publishedAt:
          type: [string, 'null']
          format: date-time
    AddPlaylistItemRequest:
      type: object
      additionalProperties: false
      required: [recordingId]
      properties:
        recordingId:
          type: string
    ReorderPlaylistItemsRequest:
      type: object
      additionalProperties: false
      required: [recordingIds]
      properties:
        recordingIds:
          type: array
          items:
            type: string
...
    SetPlaylistTagsRequest:
      type: object
      additionalProperties: false
      required: [tagIds]
      properties:
        tagIds:
          type: array
          items:
            type: string
...
    FavoritePlaylistsResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/FavoritePlaylistItem'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    FavoritePlaylistSavedResponse:
      type: object
      additionalProperties: false
      required: [id, playlistId, savedAt]
      properties:
        id:
          type: string
        playlistId:
          type: string
        savedAt:
          type: string
          format: date-time
...
    CollectionPlaylistsResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/CollectionPlaylistItem'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    CollectionPlaylistSavedResponse:
      type: object
      additionalProperties: false
      required: [id, playlistId, savedAt]
      properties:
        id:
          type: string
        playlistId:
          type: string
        savedAt:
          type: string
          format: date-time
...